refactor: 예약 가능한 시간 목록 조회 메서드 ReservationTimeRepository로 이동
