+++ src/main/java/roomescape/web/controller/ReservationTimeController.java
@@ -50,7 +50,7 @@ public class ReservationTimeController {
     @GetMapping(params = {"date", "themeId"})
     public ResponseEntity<List<BookingTimeResponseDto>> findBookable(@RequestParam("date") String date,
                                                                      @RequestParam("themeId") Long themeId) {
-        return ResponseEntity.ok(reservationTimeService.findBookable2(date, themeId));
+        return ResponseEntity.ok(reservationTimeService.findBookable(date, themeId));
     }
 
     @DeleteMapping("/{id}")
